removed simulation app
