+++ website/apps/home/views/UploadView.py
@@ -17,13 +17,13 @@ from django.db import transaction
 from django.http.response import HttpResponseBadRequest, HttpResponseRedirect
 from django.views.generic.base import TemplateView
 
-from website.apps.simulation.utils import load_simulation_file
+from website.apps.home.utils import load_simulation_file
 
 logger = logging.getLogger(__name__)
 
 
 class UploadView(TemplateView):
-    template_name = "simulation/upload.html"
+    template_name = "../templates/simulation/upload.html"
 
     @transaction.atomic
     def post(self, request, *args, **kwargs):
