--- website/apps/simulation/templates/simulation/browse.html
@@ -1,30 +1,0 @@
-{% extends "base.html" %}
-
-{% block main_content %}
-    <a href="{% url "index" %}">Back</a>
-
-    <h1>Model</h1>
-    <select id="simulation" class="form-control" style="width:350px;" >
-    {% for simulation in simulations %}
-    <option value="{{ simulation.id }}">{{ simulation.name }}</option>
-    {% endfor %}
-    </select>
-
-    <h1>Location</h1>
-    <select id="location" class="form-control" style="width:350px;" >
-    {% for location in locations %}
-    <option value="{{ location.municipality_code }}">{{ location.department }} - {{ location.municipality }}</option>
-    {% endfor %}
-    </select>
-    <a id="plot_button" class="btn btn-success" target="_blank">Generate Chart</a>
-{% endblock %}
-
-{% block extra_js %}
-    <script>
-
-        $('#plot_button').click(function () {
-           window.location.href = "/simulation/chart/" + $('#simulation').val() + "/" +  $('#location').val()
-        })
-
-    </script>
-{% endblock %}
